requirements.txt
streamlit
google-generativeai
gspread
google-auth
oauth2client
